requirements.txt
fastapi
uvicorn
ipython
instagrapi
python-multipart
Pillow
moviepy==1.0.3
pytest-asyncio
httpx
pudb
ipdb
aiofiles
requests
starlette
pytest
tinydb
python-telegram-bot
nest_asyncio
watchgod
